Add files to test a user application
Added the folder GPIO_InfiniteLedToggling which contains an example
project created using CubeMX. This project flashes the built-in LED
of the NUCLEO board.

This application will be used to test the "jump to user application"
function that will be created in the bootloader main.

This application will be built and written at memory location:
0x0800 8000, that is +32Kbytes from the start of user memory location
0x0800 0000.