Implement Timer_16 interrupt callback
Timer 16 is started in main program. The timer that waits for 5 sec
to receive a message from host to enter bootloader mode. While waiting the onboard LED
flashes slowly.

If host message is received, timer_16 is stopped and on-board LED is turned off.

Otherwise, in the timer interrupt HAL_TIM_PeriodElapsedCallback, jump to user app is performed.