Modify received_data_index variable name to current_data_index
received_data_index variable name can be misleading as it does not
really represent the index that has been received.

the name current_data_index is more appropriate as this variable depends
on the bootloader application and not on the host application.

For example.
Id message received is 106, representing index 06.
The current index that needs to be written in the Data_array can have
a different index, for example 02. When this missmatch occurs an error
frame is sent back on CAN bus with error id 2ff, data FF, received index, expected index